Machine Project 1 - RGB LED
This has the color changing LED where you can control the brightness and color cycle depending on which button interrupt is pressed. The update only happens when the button is pressed and not when the potentiometer ADC is moved. So moving the potentiometer around does not do anything UNLESS you enable the interrupt. It will start with the initial color in the color cycle at start up too.